refactor: move responsability of listing tag to repository struct
